requirements.txt
st-files-connection
gcsfs
streamlit
numpy
tensorflow
Pillow
requests
pandas
matplotlib
validators
